Router: move getProcessSummary to vote api
